Store tuple wides
