+++ test/unit/memory/CMakeLists.txt
@@ -34,5 +34,9 @@ make_unit( "unit.memory" load/unaligned/arithmetic_if_else.cpp)
 make_unit( "unit.memory" load/unaligned/logical.cpp           )
 make_unit( "unit.memory" load/unaligned/logical_if.cpp        )
 make_unit( "unit.memory" load/unaligned/logical_if_else.cpp   )
-make_unit( "unit.memory" store.cpp                            )
 make_unit( "unit.memory" compress_store.cpp                   )
+make_unit( "unit.memory" store/regular/aligned.cpp            )
+make_unit( "unit.memory" store/tuple/aligned.cpp              )
+make_unit( "unit.memory" store/regular/unaligned.cpp          )
+make_unit( "unit.memory" store/tuple/unaligned.cpp            )
+make_unit( "unit.memory" store/regular/conditional.cpp        )
